Fixed warning

+++ src/libtsduck/dtv/descriptors/tsVVCVideoDescriptor.cpp
@@ -281,11 +281,11 @@ void ts::VVCVideoDescriptor::DisplayDescriptor(TablesDisplay& disp, PSIBuffer& b
         disp << ", still pictures: " << UString::TrueFalse(buf.getBool());
         disp << ", 24-hour pictures: " << UString::TrueFalse(buf.getBool()) << std::endl;
         buf.skipBits(5);
-        const uint16_t hdr_wcg_idc = buf.getBits<uint16_t>(2);
-        disp << margin << "HDR WCG idc: " << VVCVideoDescriptor::VVCHDRandWCG(hdr_wcg_idc) << " (" << hdr_wcg_idc << ")";
+        const uint8_t hdr_wcg_idc = buf.getBits<uint8_t>(2);
+        disp << margin << "HDR WCG idc: " << VVCVideoDescriptor::VVCHDRandWCG(hdr_wcg_idc) << " (" << uint16_t(hdr_wcg_idc) << ")";
         buf.skipBits(2);
-        const uint16_t vprop_tag = buf.getBits<uint16_t>(4);
-        disp << ", video properties: " << VVCVideoDescriptor::VVCVideoProperties(hdr_wcg_idc, vprop_tag) << " (" << vprop_tag << ")" << std::endl;
+        const uint8_t vprop_tag = buf.getBits<uint8_t>(4);
+        disp << ", video properties: " << VVCVideoDescriptor::VVCVideoProperties(hdr_wcg_idc, vprop_tag) << " (" << uint16_t(vprop_tag) << ")" << std::endl;
         if (temporal && buf.canReadBytes(2)) {
             buf.skipBits(5);
             disp << margin << "Temporal id min: " << buf.getBits<uint16_t>(3);
